+++ src/modules/calendar-picker/calendar-picker.types.ts
@@ -1,0 +1,5 @@
+// Types
+import {CalendarPickerProps} from 'react-native-calendar-picker';
+
+// Props
+export interface IProps extends CalendarPickerProps {}