+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Update Person Record Component, Fixes UI and Functionalities.
 - Add SelectPicker showCustomIcon Prop
 - Add style textAlignVertical:'top' to TextInput when multiline for same behavior in both platforms
+- Add ReminderPicker Component
+- Add CalendarPicker Component
 
 --